CBCCCM-366: Implement 'Delete Form' modal dialog
