+++ cccm-frontend/src/components/RNAList.vue
@@ -303,7 +303,6 @@ export default {
       this.formTypes.unshift(this.selectedFormTypes);
     },
     async createFormAPI() {
-
       // get latest form version
       const [err, formInfo] = await getFormSummaries('CRNA', true);
       this.formData.clientNumber = "00142091";
@@ -786,30 +785,21 @@ export default {
       if (formID != null) {
         formID = formID.toString();
       }
-      if (formType === 'SARA') {
-        this.$router.push({
-          name: 'saracmp',
-          params: {
-            formID: formID,
-            csNumber: this.clientNum
-          }
-        });
-      } else if (formType === 'CRNA') {
-        let linkedSara = false;
+      let linkedSara = false;
+      if (formType === 'CRNA') {
         if (relatedFormTypeId != null) {
           linkedSara = true;
         }
-        this.$router.push({
-          name: 'crnacmp',
-          params: {
-            formID: formID,
-            csNumber: this.clientNum,
-            linkedSara: linkedSara
-          }
-        });
-      } else {
-        console.error("Form type not supported");
       }
+      this.$router.push({
+        name: "cmpform",
+        params: {
+          formType: formType,
+          formID: this.newCreatedFormId,
+          csNumber: this.clientNum,
+          linkedSara: linkedSara
+        }
+      });
     },
     async formClone(formID) {
       console.log("formClone", formID);
@@ -823,24 +813,19 @@ export default {
       this.createFormAPI();
       //Redirect User to the newly created form
       console.log("newCreatedFormID: ", this.newCreatedFormId);
+      let formType = "CRNA";
       if (this.selectedFormTypeValue.includes("sara")) {
-        this.$router.push({
-          name: "saracmp",
-          params: {
-            formID: this.newCreatedFormId,
-            csNumber: this.clientNum
-          }
-        });
-      } else {
-        this.$router.push({
-          name: "crnacmp",
-          params: {
-            formID: this.newCreatedFormId,
-            csNumber: this.clientNum,
-            linkedSara: false
-          }
-        });
+        formType = "SARA";
       }
+      this.$router.push({
+        name: "cmpform",
+        params: {
+          formType: formType,
+          formID: this.newCreatedFormId,
+          csNumber: this.clientNum,
+          linkedSara: false
+        }
+      });
     },
     formCreate() {
       console.log("Create form btn click");
