+++ cccm-frontend/src/components/sara-cmp/saraCmpForm.vue
@@ -1,57 +1,5 @@
 <template>
   <div data-app class="sara-cmp-form">
-    <!--Form delete modal dialog -->
-    <v-btn
-      id="id_modal_deleteForm"
-      v-show=false
-      @click.stop="dialog = true"
-    ></v-btn>
-    <v-dialog
-        v-model="dialog"
-        persistent
-        max-width="550"
-      >
-      <v-card>
-        <div class="col-sm-12 m-7">
-          <v-card-title >
-            Select what you want to delete:
-          </v-card-title>
-          <v-checkbox
-            v-model="selectedFormTypeValue"
-            label="CRNA-CMP"
-            value="crna"
-          ></v-checkbox>
-          <v-checkbox
-            v-model="selectedFormTypeValue"
-            :readonly=true
-            label="SARA-CMP"
-            value="sara"
-          ></v-checkbox>
-          <v-card-title>
-          Are you sure you want to delete?
-        </v-card-title>
-        <v-card-text>
-          The form(s) and all the information you have entered will be deleted and you will be directed to the client's RNA list. 
-        </v-card-text>
-        </div>
-        
-        <v-card-actions>
-          <v-btn
-            @click="dialog = false"
-          >
-          No, I don't want to delete
-          </v-btn>
-          <v-spacer></v-spacer>
-          <v-btn
-            color="#f81e41"
-            dark
-            @click="handleDeleteFormBtnClick"
-          >
-            Yes, delete form(s)
-          </v-btn>
-        </v-card-actions>
-      </v-card>
-    </v-dialog>
     <div class="main">
       <div class="wrap">
         <div class="mainRow">
@@ -117,7 +65,6 @@ export default {
       data_buttonGroup: {},
       initData: {},
       dataMap: {},
-      dialog: false,
       baseURL: import.meta.env.BASE_URL,
       clientNum: '',
       formId: '',
@@ -199,24 +146,7 @@ export default {
       console.log('Print Form.')
     },
     handleCancelForm() {
-      console.log("Cancel Form");
-      let modal = document.getElementById("id_modal_deleteForm");
-      if (modal != null) {
-        modal.click();
-      }
-    },
-    handleDeleteFormBtnClick() {
-      this.dialog = false;
-      //this.deleteForm();
-
-      //Redirect User back to clientRecord.RNAList
-      this.$router.push({
-        name: 'clientrecord',
-        params: {
-          clientNum: this.clientNum,
-          tabIndex: 'tab-rl'
-        }
-      });
+      this.$emit('cancelFormClicked');
     },
     handleNavChildCallback(parentNavCurLocationFromChild) {
       this.parentNavCurLocation = parentNavCurLocationFromChild;
